src/mcp-web-search/pyproject.toml
[project]
name = "mcp-web-search"
version = "0.1.0"
description = "MCP web search server — fetch/scrape (dev) or Bing Grounding (prod)"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "beautifulsoup4>=4.12",
    "starlette>=0.37",
    "uvicorn>=0.29",
]

[project.optional-dependencies]
prod = [
    "azure-identity>=1.19.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
markers = ["integration: requires external API access"]
addopts = "-m 'not integration'"
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["mcp_web_search"]
